Python/requirements.txt
pycryptodome
cx_Oracle==6.3.1
pandas==0.23.1
swifter==0.153